Move interaces to a separate file and  remove restrictions on letters being typed in boxes that leave gaps in the guess
